Refactored: cache in separate source to avoid circular imports
